+++ mod05/ex01/Form.hpp
@@ -18,8 +18,8 @@ class Form {
 		
 		std::string		getName(void) const;
 		bool			getIsSigned(void) const;
-		unsigned short	getRequiredGradeToSign(void) const;
-		unsigned short	getRequiredGradeToExecute(void) const;
+		size_t	getRequiredGradeToSign(void) const;
+		size_t	getRequiredGradeToExecute(void) const;
 
 		void			beSigned(const Bureaucrat &b);
 
@@ -41,8 +41,8 @@ class Form {
 	private:
 		const std::string			_name;
 		bool						_isSigned;
-		const unsigned short		_requiredGradeToExecute;
-		const unsigned short		_requiredGradeToSign;
+		const size_t		_requiredGradeToExecute;
+		const size_t		_requiredGradeToSign;
 };
 
 std::ostream &operator<<(std::ostream& os, const Form& bureaucrat);
